Cooking Agent/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
boto3==1.35.82
botocore==1.35.82
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
gitdb==4.0.11
GitPython==3.1.43
groq==0.13.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jmespath==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
phidata==2.7.3
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==308
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
s3transfer==0.10.4
setuptools==75.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.2.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wheel==0.44.0
